moved properlyCapitalizeInitiator to its own file and add tests
