+++ src/components/CRAndTDLSidePanel/_utils.ts
@@ -1,0 +1,13 @@
+const properlyCapitalizeInitiator = (fullName: string) => {
+  const [namePart, parenthesisPart] = fullName.split(' (');
+
+  const capitalizedName = namePart
+    .toLowerCase()
+    .replace(/\b\w/g, char => char.toUpperCase());
+
+  return parenthesisPart
+    ? `${capitalizedName} (${parenthesisPart}`
+    : capitalizedName;
+};
+
+export default properlyCapitalizeInitiator;
